Separate unwind information helpers
